Refactor methods for sending peer messages
